Better teardown process.
LSF's bkill sends a TERM and then a KILL moments later if TERM is not
fulfilled within a time limit.  The slave processes do not always respond to
TERM quickly enough to execute a teardown, which means that cleanup rarely
happens in a bkill situation.  This approach makes LSF itself perform the
cleanup.